Boolean if the rights to use can be transferable
Licenses can by default be assigned or transferred to another party, unless it is specifically restricted in the license agreement. A company who wants to sell their assets (since an exclusive license may be a significant asset) will want the ability to transfer, but the licensor may not want to do business with a new person that they didn’t approve.